pyproject.toml
[project]
name = "expert-selector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.66.0",
    "pandas>=3.0.1",
    "pyqt5>=5.15.11",
    "python-dotenv>=1.0.1",
]
